# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Gacha Game Backend API
  version: 1.0.0
  description: API for managing gacha rolls, user inventories, auctions, transactions, and user authentication.
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /users/auth/login:
    post:
      tags:
        - PLAYER
      summary: Login a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: player123
                password:
                  type: string
                  example: securepassword
      responses:
        '200':
          description: Successful login
        '401':
          description: Invalid credentials
        '500':
          description: Database error
        '501':
          description: Unexpected error
  /users/auth/register:
    post:
      tags:
        - PLAYER
      summary: Register a new player
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: new_player123
                password:
                  type: string
                  example: securepassword
                email:
                  type: string
                  example: new_player@example.com
                image:
                  type: object
      responses:
        '200':
          description: Player successfully registered
        '400':
          description: Missing required fields
        '409':
          description: Username or email already exists
        '500':
          description: Unexpected error
  /users/auth/logout:
    delete:
      tags:
        - PLAYER
      summary: Logout a player
      responses:
        '200':
          description: Successful logout
        '401':
          description: Invalid user type
        '500':
          description: Unexpected error        
  /users/auth/delete:
    delete:
      tags:
        - PLAYER
      summary: Delete a user account
      responses:
        '200':
          description: Account deleted successfully
        '400':
          description: Session token is required
        '408':
          description: Session token not found
        '500':
          description: Unexpected error
  /users/player/update:
    put:
      tags:
        - PLAYER
      summary: Update player account details
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: updated_username
                password:
                  type: string
                  example: updated_password
                email:
                  type: string
                  example: updated_email@example.com
                image:
                  type: object           
      responses:
        '200':
          description: Account successfully updated
        '400':
          description: Invalid user type
        '408':
          description: User not found
        '500':
          description: Unexpected error
  /users/player/my_gacha_list:
    get:
      tags:
        - PLAYER
      summary: Retrieve the list of gacha items for a user
      responses:
        '200':
          description: List of gacha items retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    gacha_id:
                      type: string
                      example: "1"
                    name:
                      type: string
                      example: "Auto"
                    rarity:
                      type: string
                      example: "epic"
                    status:
                      type: string
                      example: "available"
                    description:
                      type: string
                      example: "Fast auto"
                    image:
                      type: string
                      format: base64
                      example: "iVBORw0KGgoAAAANSUhEUgAA..."
                    locked:
                      type: string
                      example: "unlocked"
                    acquired_date:
                      type: string
                      example: "2023-10-01T12:00:00Z"
        '404':
          description: No gacha items found for the user
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No gacha items found for the user"
  /users/player/gacha/{user_id}/{gacha_id}:
    get:
      tags:
        - PLAYER
      summary: Retrieve information about a specific gacha item for a user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
          example: "1"
        - in: path
          name: gacha_id
          required: true
          schema:
            type: string
          example: "1"
      responses:
        '200':
          description: Gacha item information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  gacha_id:
                    type: string
                    example: "1"
                  name:
                    type: string
                    example: "Auto"
                  rarity:
                    type: string
                    example: "epic"
                  status:
                    type: string
                    example: "available"
                  description:
                    type: string
                    example: "Fast auto"
                  image:
                    type: string
                    format: base64
                    example: "iVBORw0KGgoAAAANSUhEUgAA..."
        '404':
          description: User not found or gacha item not found
  /users/player/real_money_transaction:
    post:
      tags:
        - PLAYER
      summary: Handle real money transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  example: 100.0
      responses:
        '200':
          description: Account topped up successfully
        '400':
          description: Missing amount for transaction or invalid amount
  /users/player/get_user_balance:
    get:
      tags:
        - PLAYER
      summary: Retrieve the balance of a user
      responses:
        '200':
          description: User balance retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  currency_balance:
                    type: number
                    example: 1500.0
        '400':
          description: Invalid user type
        '408':
          description: User not found
        '500':
          description: Unexpected error
  /users/player/get_user:
    get:
      tags:
        - PLAYER
      summary: Get user information
      description: Retrieve information for the authenticated user. Only users with the 'PLAYER' type can access this endpoint.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                  username:
                    type: string
                  email:
                    type: string
                  currency_balance:
                    type: number
                  image:
                    type: object
                  session_token:
                    type: string
        '404':
          description: User not found
        '500':
          description: Unexpected error
  /users/player/get_user/{user_id}:
    get:
      tags:
        - PLAYER
      summary: Get user information by ID (Admin only)
      description: >-
        Retrieve user information by ID. This endpoint is restricted to admin users 
        and is hidden from the API documentation.
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user to retrieve.
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                  username:
                    type: string
                  email:
                    type: string
                  currency_balance:
                    type: number
                  image:
                    type: object
                  session_token:
                    type: string
        '404':
          description: User not found
        '500':
          description: Unexpected error
  /users/player/update_balance/{user_type}:
    put:
      tags:
        - PLAYER
      summary: Update the balance of a user
      parameters:
        - in: path
          name: user_type
          required: true
          schema:
            type: string
            enum: [PLAYER, ADMIN]
          example: "PLAYER"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  example: "1"
                amount:
                  type: string
                  example: "100"
                type:
                  type: string
                  example: "credit"
      responses:
        '200':
          description: Balance updated successfully
        '400':
          description: Missing data to update balance
        '404':
          description: User not found
        '500':
          description: Unexpected error
...